requirements.txt
openvino-dev
transformers
numpy
streamlit
opencv-python
